Add a variable that determines whether the role will be read-only or not
By the default the value if true, indicating that the role should be
read-only.  This agrees with the way the role functioned previously,
so the change should cause no surprises.

If the value is set to false then the role created will allow for
writing in addition to reading.

Note that this new value is passed to cisagov/s3-read-role-tf-module
in read_terraform_state_role.tf.